+++ src/briefy/leica/tools/dump_events.py
@@ -12,12 +12,16 @@ from briefy.leica.models.professional.workflows import ProfessionalWorkflow
 from briefy.leica.models.professional.workflows import LocationWorkflow
 from briefy.leica.models.professional.workflows import LinkWorkflow
 from briefy.leica.models.project.workflows import ProjectWorkflow
+from briefy.leica.models.user.workflows import BriefyUserProfileWorkflow
+from briefy.leica.models.user.workflows import CustomerUserProfileWorkflow
 
 entities = [
     'asset',
     'assignment',
+    'briefyuserprofile',
     'comment',
     'customer',
+    'customeruserprofile',
     'order',
     'pool',
     'professional',
@@ -44,7 +48,9 @@ workflows = [
     ProfessionalWorkflow,
     LocationWorkflow,
     LinkWorkflow,
-    ProjectWorkflow
+    ProjectWorkflow,
+    BriefyUserProfileWorkflow,
+    CustomerUserProfileWorkflow,
 ]