+++ src/briefy/leica/worker/actions.py
@@ -11,14 +11,14 @@ def validate_assignment(laure_data: object) -> (bool, dict):
     """Perform necessary transition after photo set validation.
 
     :param laure_data: Python object representing Laure data after assignment validation
-    :return: Flag indicating wether operation was successfull, empty dict
+    :return: Flag indicating whether operation was sucessful, empty dict
     """
     assignment_id = laure_data.assignment_info.id
     assignment = Assignment.query().get(assignment_id)
 
     if assignment.state != 'asset_validation':
         logger.error(
-            '''Got message to validade assignemnt '{0}' which is in state '{1}' '''.format(
+            '''Got message to validate assignment '{0}' which is in state '{1}' '''.format(
                 assignment_id,
                 assignment.state
             )
@@ -30,7 +30,7 @@ def validate_assignment(laure_data: object) -> (bool, dict):
             assignment_id
         )
     )
-    assignment.workflow_context = SystemUser
+    assignment.workflow.context = SystemUser
     with transaction.manager:
         assignment.workflow.validate_assets(
             message='Validated submission.'
@@ -43,7 +43,7 @@ def invalidate_assignment(laure_data: object) -> (bool, dict):
     """Perform necessary transition and field update after photo set was deemed invalid.
 
     :param laure_data: Python object representing Laure data after assignment validation
-    :return: Flag indicating wether operation was successfull, empty dict
+    :return: Flag indicating whether operation was successful, empty dict
     """
     assignment_id = laure_data.assignment_info.id
     assignment = Assignment.query().get(assignment_id)
@@ -60,10 +60,12 @@ def invalidate_assignment(laure_data: object) -> (bool, dict):
         laure_data.validation_info.complete_feedback
     )
 
-    feedback_comment = Comment(author_id=SystemUser.id,
-                               author_role='system',
-                               to_role='professional',
-                               content=feedback_text)
+    feedback_comment = Comment(
+        author_id=SystemUser.id,
+        author_role='system',
+        to_role='professional',
+        content=feedback_text
+    )
 
     logger.info(
         '''Assignment '{0}' assets reported as not sufficient. Transitioning back '''
@@ -89,23 +91,31 @@ def approve_assignment(laure_data: object) -> (bool, dict):
     :param laure_data: Python object representing Laure data after assignment approval
     :return: Flag indicating wether operation was successfull, empty dict
     """
-    assignment_id = laure_data.assignment_info.id
+    assignment_id = laure_data.guid
     assignment = Assignment.query().get(assignment_id)
 
     # Ensure the correct key is updated and object is set as dirty
-    delivery_info = assignment.order.delivery
+    delivery_info = assignment.order.delivery or {}
     delivery_info['gdrive'] = laure_data.delivery_url
+    delivery_info['archive'] = laure_data.archive_url
 
-    logger.info('''Assets copied over on laure - commiting delivery URL to order '{}' '''.format(
-                assignment.order.id))
+    logger.info(
+        '''Assets copied over on laure - committing delivery URL to order '{order_id}' '''.format(
+            order_id=assignment.order.id
+        )
+    )
 
     with transaction.manager:
         assignment.order.delivery = delivery_info
 
     # TODO: Unless google drive is phased out soon, this URL should be
     # stored on the model.
-    logger.info('''Assets for assignment '{}' are archived at '{}' '''.format(
-                assignment_id, laure_data.archive_url))
+    logger.info(
+        '''Assets for assignment '{0}' are archived at '{1}' '''.format(
+            assignment_id,
+            laure_data.archive_url
+        )
+    )
 
     return True, {}
 
@@ -114,28 +124,35 @@ def asset_copy_malfunction(laure_data: object) -> (bool, dict):
     """Perform necessary transition and field update after photo set was deemed invalid.
 
     :param laure_data: Python object representing Laure data after assignment validation
-    :return: Flag indicating wether operation was successfull, empty dict
+    :return: Flag indicating whether operation was successful, empty dict
     """
     assignment_id = laure_data.assignment_info.id
     assignment = Assignment.query().get(assignment_id)
 
     if assignment.state != 'asset_validation':
-        logger.error('''Got message to invalidate '{}' which is in state '{}' '''.format(
-                    assignment_id, assignment.state))
+        logger.error(
+            '''Got message to invalidate '{0}' which is in state '{1}' '''.format(
+                assignment_id,
+                assignment.state
+            )
+        )
         return False, {}
 
     text = ('''This assignment's assets could not be copied automatically!\n'''
-            '''Please take manual actions that may be needed. ''')
+            '''Please take manual actions that may be needed.''')
 
     comment = Comment(
         author_id=SystemUser.id,
         author_role='system',
         to_role='qa_manager',
-        content=text
+        content=text,
+        internal=True,
     )
 
-    logger.warn('''There was a problem copying assets to delivery folders.'''
-                '''Adding comment to assignment'''.format(assignment_id))
+    logger.warn(
+        '''There was a problem copying assets to delivery folders.'''
+        '''Adding comment to assignment {0}'''.format(assignment_id)
+    )
 
     comment.workflow_context = SystemUser
     assignment.workflow_context = SystemUser