+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
     * Split workflows for Briefy and Customer profiles (ericof).
     * Fix Submission Date calculation on Assignment (ericof).
     * Expose initial password on UserProfile creation (ericof).
+    * Set timezone on new and updated OrderLocations (ericof).
 
 
 2.0.3 (2017-02-14)