+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
     * Fix Submission Date calculation on Assignment (ericof).
     * Expose initial password on UserProfile creation (ericof).
     * Set timezone on new and updated OrderLocations (ericof).
+    * Fix remove_availability transition: now the new assignment is created after cancel the old one (rudaporto).
 
 
 2.0.3 (2017-02-14)